+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Thanks to: @dathbe.
 - [clock] Fixed missing icons when no other modules with icons is loaded (#3834)
 - [weather] Fixed handling of empty values in weathergov providers handling of precipitationAmount (#3859)
 - [calendar] Fix regression handling of limit days (#3840)
+- [calendar] Fixed regression of calendarfetcherutils.shouldEventBeExcluded (#3841)
 
 ## [2.32.0] - 2025-07-01